backend/requirements.txt
fastapi
uvicorn
pymongo
langchain
langgraph
python-dotenv
pinecone
groq
pydantic
langchain-groq
python-multipart
certifi